+++ hosts/upsilon/darwin-configuration.nix
@@ -1,3 +1,7 @@
-{ ... }: {
+{ lib, ... }: {
   networking.knownNetworkServices = [ "Wi-Fi" ];
+
+  services.synergy.client.enable = true;
+  services.synergy.client.serverAddress = "phi-nixos.local";
+  services.synergy.client.tls = true;
 }